+++ policyengine_us/variables/gov/states/ut/dwf/ccap/ut_child_care_subsidies.py
@@ -1,0 +1,11 @@
+from policyengine_us.model_api import *
+
+
+class ut_child_care_subsidies(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = "Utah child care subsidies"
+    unit = USD
+    definition_period = YEAR
+    defined_for = StateCode.UT
+    adds = ["ut_ccap"]